AdvancingX Instagram Data Analysis
This document outlines a data analysis conducted using content performance data from AdvancingX's Instagram page. The objective of this analysis was to deliver strategic recommendations for marketing campaign ideas that could enhance AdvancingX's leveraging of Instagram. For this project, I utilized Excel and SQL for data cleaning and interpretation, while Tableau was employed for data visualization.